feat(workspace): 5 minute timeout because codex is slow
Eventually, this will be replaced by a more accurate check for signs of
life, meaning when the LLM has signed its last token but for right now a
five minute timeout makes sense and hey if you're running a local model
this will also be useful because local models are really slow if you
weren't aware of that. I should probably not be dictating commit
messages like this, but it's kind of fun.